# kubegram-server/docs/swagger.yaml
basePath: /api/public/v1
definitions:
...
  github_com_kubegram_kubegram-server_src_models.User:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      provider:
        description: OAuth linkage
        type: string
      provider_id:
        type: string
      role:
        description: '"admin" | "manager" | "team_member"'
        type: string
      teamID:
        type: integer
      updated_at:
        type: string
      updatedAt:
        type: string
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  github_com_kubegram_kubegram-server_src_models.Provider:
    properties:
      id:
        type: integer
      provider:
        type: string
      apiUrl:
        type: string
      models:
        items:
          type: string
        type: array
    type: object
  github_com_kubegram_kubegram-server_src_models.ProvidersList:
    properties:
      providers:
        items:
          $ref: '#/definitions/github_com_kubegram_kubegram-server_src_models.Provider'
        type: array
    type: object
host: localhost:8090
securityDefinitions:
  OAuth2:
    type: oauth2
    flow: accessCode
    authorizationUrl: /auth/{provider}/login
    tokenUrl: /auth/{provider}/callback
    scopes:
      admin: Admin access for internal operations
      read: Read access
      write: Write access
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for Kubegram application
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Kubegram API
  version: "1.0"
paths:
...
  /api/v1/providers:
    get:
      description: Get a list of enabled LLM providers for the user
      security:
      - OAuth2:
        - read
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_kubegram_kubegram-server_src_models.ProvidersList'
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: Get enabled providers
      tags:
      - providers
    post:
      description: Create a new provider configuration
      security:
      - OAuth2:
        - write
      consumes:
      - application/json
      parameters:
      - in: body
        name: body
        required: true
        schema:
          type: object
          properties:
            provider:
              type: string
            apiUrl:
              type: string
            models:
              type: array
              items:
                type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_kubegram_kubegram-server_src_models.Provider'
        "400":
          description: Bad Request
        "404":
            description: Company Not Found
        "500":
          description: Internal Server Error
      summary: Add a provider
      tags:
      - providers
  /api/v1/providers/{id}:
    put:
      description: Update a provider configuration
      security:
      - OAuth2:
        - write
      consumes:
      - application/json
      parameters:
      - in: path
        name: id
        required: true
        type: integer
      - in: body
        name: body
        required: true
        schema:
          type: object
          properties:
            apiUrl:
              type: string
            models:
              type: array
              items:
                type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_kubegram_kubegram-server_src_models.Provider'
        "400":
          description: Bad Request
        "403":
            description: Unauthorized
        "404":
            description: Not Found
        "500":
          description: Internal Server Error
      summary: Update a provider
      tags:
      - providers
    delete:
      description: Delete a provider
      security:
      - OAuth2:
        - write
      parameters:
      - in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "403":
            description: Unauthorized
        "404":
            description: Not Found
        "500":
          description: Internal Server Error
      summary: Delete a provider
      tags:
      - providers
...
  /api/v1/users/me:
    get:
      description: Get current authenticated user
      security:
      - OAuth2:
        - read
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_kubegram_kubegram-server_src_models.User'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get current user
      tags:
      - users
  /api/v1/users:
    get:
      description: Get a list of all registered users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_kubegram_kubegram-server_src_models.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: User object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/github_com_kubegram_kubegram-server_src_models.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_kubegram_kubegram-server_src_models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a user
      tags:
      - users
...
swagger: "2.0"